#52 add pagination

+++ server/src/main/java/com/example/pollprojectmain/service/Impl/UserDetailsImpl.java
@@ -20,14 +20,14 @@ public class UserDetailsImpl implements UserDetails {
 
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private Integer id;
     private String username;
     private String email;
     @JsonIgnore
     private String password;
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserDetailsImpl(Long id, String username, String email, String password,
+    public UserDetailsImpl(Integer id, String username, String email, String password,
                            Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.username = username;
@@ -41,7 +41,7 @@ public class UserDetailsImpl implements UserDetails {
         authorities.add(new SimpleGrantedAuthority("ROLE_" + user.getRole().toString()));
 
         return new UserDetailsImpl(
-                user.getId().longValue(),
+                user.getId(),
                 user.getUsername(),
                 user.getEmail(),
                 user.getPassword(),
@@ -63,7 +63,7 @@ public class UserDetailsImpl implements UserDetails {
         return username;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
